proj-titanic-survivor-prediction/pyproject.toml
[tool.poetry]
name = "proj-titanic-survivor-prediction"
version = "0.1.0"
description = ""
authors = ["Jukkraput <78992646+jukkraputp@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
pandas = "^2.2.1"
sqlalchemy = "1.4.28"
pymysql = "^1.1.0"
apache-airflow = {extras = ["gcp"], version = "^2.8.1"}
apache-airflow-providers-mysql = "^5.5.3"
google-cloud = "^0.34.0"
google-api-python-client = "^2.119.0"
google-auth = "^2.28.1"
google-auth-httplib2 = "^0.2.0"
google-cloud-bigquery = "^3.17.2"
apache-airflow-providers-google = "^10.15.0"
streamlit = "^1.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
